requirements.txt
growwapi
pyotp
pandas
numpy
pandas_ta
pydantic>=2.7
fastapi
uvicorn[standard]
pyyaml
requests
python-dateutil
